fixed goals pathway missing in App.jsx

+++ backend/src/main/java/edu/cit/fitflow/service/UserService.java
@@ -35,7 +35,7 @@ public class UserService {
 	}
 
   // Added method to fetch user by userId
-  public Optional<UserEntity> getUserById(Long userId) {
+  public Optional<UserEntity> getUserById(Integer userId) {
     return urepo.findById(userId);
   }
